+++ tests/Checkout/ExpressCheckout/Service/ExpressCustomerServiceTest.php
@@ -14,6 +14,7 @@ use Shopware\Core\Checkout\Customer\SalesChannel\AccountService;
 use Shopware\Core\Checkout\Customer\SalesChannel\RegisterRoute;
 use Shopware\Core\Framework\DataAbstractionLayer\EntityRepository;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
+use Shopware\Core\Framework\Validation\DataBag\RequestDataBag;
 use Shopware\Core\System\SalesChannel\Context\SalesChannelContextService;
 use Shopware\Core\System\SalesChannel\Context\SalesChannelContextServiceParameters;
 use Shopware\Core\Test\TestDefaults;
@@ -115,7 +116,7 @@ class ExpressCustomerServiceTest extends TestCase
 
     private function doLogin(Order $order): CustomerEntity
     {
-        $contextToken = $this->createCustomerService()->loginCustomer($order, $this->getSalesChannelContext());
+        $contextToken = $this->createCustomerService()->loginCustomer($order, $this->getSalesChannelContext(), new RequestDataBag());
 
         $context = $this->getContainer()->get(SalesChannelContextService::class)->get(
             new SalesChannelContextServiceParameters(